Get urgency theory to build (omitting a "net" theorem)
There is a theorem at the bottom of TCP1_urgencyScript.sml whose
statement cannot even parse given the way that script file is set up.
The required ancestry would be in TCP1_net* theories, but those theories
couldn't build either because they don't properly define the syntax they
are using, with an -- _ --=> arrow.

I'm going to assume that stuff is a later accretion that isn't supposed
to work, and have commented out the addition to the urgency script.
Now, when you type Holmake, it runs to successful completion.  The
generated executables are almost certainly nonsense however.