Reinstate time-elapse reduction arrow as -- .. --=>
This relation sits just above the non-time-aware reduction
relation (which is -- .. -->), and pushes time along in conjunction with
what it is to be an urgent transition at the lower level.

This makes the theorem in TCP1_urgency make sense again (so it can be
uncommented). Reinstatement was partly informed by files on an old
machine with a 2 July 2007 time-stamp. The seemingly conflicting
theories (TCP1_net1 and TCP1_net1_to_net) are not used anywhere, and
modern Holmake won't even attempt to build them.

If the CVS repository is still available (at zonule.cl.cam.ac.uk?), it
might be interesting to look at the history to see why the syntax was
adjusted.