allow for sock.is2 = NONE and sock.ps2 = NONE in a trace record
Consider a symbolic state which has a socket in LISTEN state, on some is1/ps1.

Now, a remote host sends an ACK to that socket.  This will certainly be dropped
(with a RST), deliver_in_1b fires.  While the symbolic state is not modified,
there may still be a Lh_trace (of kind TA_OUTPUT) with a full quad be generated
by the OS (as it is atm by FreeBSD-CURRENT).  This Lh_trace will not pass
trace_2 since the symbolic state has is2/ps2 = NONE.

With this change, the Lh_trace is accepted under these conditions.  Not entirely
sure whether this introduces regression (the model now accepts more traces than
before) - need to revisit at some point, but the trace
(rcv-ack-without-data-listen) passes now.