fix simultaneours open (deliver_in_2)
FreeBSD does _not_ call cc_conn_init yet, but the model does (and sets the
congestion window).  The model then increases this value in deliver_in_3
(when transitioning to ESTABLISHED) by t_maxseg, which leads to
14600 + 1460 = 16060 in the model, while FreeBSD only calls cc_conn_init
once transitioning to ESTABLISHED (and thus having 14600 there).

This patch is a HACK, but in deliver_in_3 (where SYN_RCVD -> ESTABLISHED) we
lost the predicate whether we did an active or simultaneous open.

In addition, FreeBSD these days sends a SYN+ACK as reply to SYN for the
simultaneous open (used to send ACK only).  In the SYN+ACK, timestamp and
window_scale options are both present (in contrast to 1323!?), even if the
SYN from the other party didn't contain window scaling or timestamp.