Avoid first received FIN+ACK to be treat as duplicate ACK (simultaneous close)
I still lack a clean definition of what a duplicate ACK is.  According
to Stevens Vol.2 (conjunction of the following 5 things):
 - ack is <= snd_una
 - LENGTH of received segment is 0
 - advertised window has not changed
 - rexmit timer is non-zero (outstanding transmission)
 - segment contains the biggest ack we've seen (= snd_una)

FreeBSD earlier did not handle the first FIN+ACK if the ACK was a duplicate.
This lead to situations where the FIN was never processed.  They fixed this
in rev239672 (commit message includes: if we don't do that, FIN is never
processed, and both sides loop sending each other FIN+ACK segments (retransmit)
until timeout occurs).

It seems sensible to _process_ the first FIN, but I'd argue it is also sensible
to count it as a duplicate ACK - it does not acknowledge any new data.  But then
again it seems to be ok to send FIN+ACK when close() (or shutdown()) was called
on the socket.