deliver_in_8: check for seq being in window
only if seq is in the window, we are supposed to RST the SYN.  otherwise we are
supposed to send an ACK to the other side.  This is part of RFC793 again,
reiterated in 5961 sec 4.2 (which again suggests to be more picky about which
SYN segments to reply with RST).  the model implements only 793, not 5961.