rcv_wnd update:  streamline with how FreeBSD handles receive window updates
specifically, intuitively "rcv_nxt + rcv_wnd = rcv_adv" should be true.

FTR Stevens Vol2 doesn't talk much about rcv_adv

but it isn't, FreeBSD uses rcv_wnd only in tcp_input, where first it is assigned
to: rcv_wnd = imax(imax(sbspace(so->so_rcv), 0), rcv_adv - rcv_nxt) thus, it is
rather a local variable than a stateful variable (tcp_output uses in similar way
the local variable recwin).

rcv_adv is only assigned to in tcp_output (and its new value is rcv_nxt + recwin
if this is greater than the old one - where recwin is massaged to not shrink and
don't be a silly window).

rcv_nxt is - as described in stevens - assigned to in tcp_input (for received
syn/fin, and actual data)

so, why this change?  because if we adjust rcv_wnd and rcv_adv in tcp_output, we
may update the rcv_wnd earlier than FreeBSD, and our rcv_adv (which is used for
evaluating whether or not an incoming segment is in-window or out of it)

NB: this change is slightly incomplete (specifically, there are at two places
comments about the non-symmetric rcv_wnd updates between the model and the
FreeBSD implementation), to be revised (and force-pushed) once trace checking
succeeds.